Moved definition of LinkClock into .hpp
