+++ lang/LangPrimSource/PyrSched.cpp
@@ -50,6 +50,7 @@
 #include "SC_Lock.h"
 
 #include "SC_Clock.hpp"
+#include "SC_Ableton_Link.hpp"
 
 #include <boost/sync/semaphore.hpp>
 #include <boost/sync/support/std_chrono.hpp>
@@ -233,12 +234,6 @@ using monotonic_clock = std::conditional<std::chrono::high_resolution_clock::is_
 
 static std::chrono::high_resolution_clock::time_point hrTimeOfInitialization;
 
-#ifdef SC_ABLETON_LINK
-  class LinkClock {
-    public:
-      static void Init();
-  };
-#endif
 SCLANG_DLLEXPORT_C void schedInit()
 {
 	using namespace std::chrono;
